chistes/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
codegen==1.0
Django==4.0.4
djangorestframework==3.13.1
idna==3.3
patterns==0.3
python-dateutil==2.8.2
pytz==2022.1
pyxattr==0.7.2
requests==2.27.1
routers==0.10.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
utils2==0.1.2
views==0.3
